Removing CLI pkg build - this has been a problematic element of our build for a while now and it is rarely utilised - rather than continuing to maintain this, getting rid of it.
