+++ packages/cli/.gitignore
@@ -5,7 +5,6 @@ build/
 docker-error.log
 envoy.yaml
 *.tar.gz
-prebuilds/
 dist/
 budibase-automation/
 budibase-component/
