+++ packages/cli/src/index.ts
@@ -1,10 +1,9 @@
 #!/usr/bin/env node
 process.env.DISABLE_PINO_LOGGER = "1"
-import "./prebuilds"
 import "./environment"
 import { getCommands } from "./options"
 import { Command } from "commander"
-import { getHelpDescription } from "./utils"
+import { getHelpDescription, error } from "./utils"
 import { version } from "../package.json"
 
 // add hosting config
@@ -20,6 +19,23 @@ async function init() {
   // this will stop the program if no command found
   await program.parseAsync(process.argv)
 }
+
+const events = ["exit", "SIGINT", "SIGUSR1", "SIGUSR2", "uncaughtException"]
+events.forEach(event => {
+  process.on(event, (evt?: number) => {
+    if (evt && !isNaN(evt)) {
+      return
+    }
+    if (evt) {
+      console.error(
+        error(
+          "Failed to run CLI command - please report with the following message:"
+        )
+      )
+      console.error(error(evt))
+    }
+  })
+})
 
 init().catch(err => {
   console.error(`Unexpected error - `, err)
